chore: move button position and size
