+++ src/components/Canary/index.tsx
@@ -1,26 +1,21 @@
 import { HealthChecksResponse } from "@flanksource-ui/api/types/health";
 import { isCanaryUI } from "@flanksource-ui/context/Environment";
 import { useHealthPageContext } from "@flanksource-ui/context/HealthPageContext";
-import { useUserAccessStateContext } from "@flanksource-ui/context/UserAccessContext/UserAccessContext";
-import { Button } from "@flanksource-ui/ui/Buttons/Button";
 import HealthPageSkeletonLoader from "@flanksource-ui/ui/SkeletonLoader/HealthPageSkeletonLoader";
 import clsx from "clsx";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { FaFilter } from "react-icons/fa";
-import { VscDebugRerun } from "react-icons/vsc";
 import { useSearchParams } from "react-router-dom";
 import { Badge } from "..";
 import useRefreshRateFromLocalStorage from "../../hooks/useRefreshRateFromLocalStorage";
 import BulkCheckRunBar from "./BulkCheckRun/BulkCheckRunBar";
-import BulkCheckRunContext from "./BulkCheckRun/BulkCheckRunContext";
+import { useBulkCheckRun } from "./BulkCheckRun/BulkCheckRunContext";
 import CanaryFiltersBar from "./CanaryFilters/TopBar/CanaryFiltersBar";
 import { CanaryInterfaceMinimal } from "./CanaryInterface";
 import CanarySidebar from "./Sidebar/CanarySidebar";
 import { isHealthy } from "./filter";
 import { decodeUrlSearchParams } from "./url";
 import { useHealthUserSettings } from "./useHealthUserSettings";
-
-const bulkRunAllowedRoles = ["admin", "editor", "responder", "commander"];
 
 const getPassingCount = (checks: any) => {
   let count = 0;
@@ -74,54 +69,7 @@ export function Canary({
   }, [hidePassing, searchParams]);
 
   const [isLoading, setIsLoading] = useState(true);
-  const [isBulkRunMode, setIsBulkRunMode] = useState(false);
-  const [selectedCheckIds, setSelectedCheckIds] = useState<Set<string>>(
-    new Set()
-  );
-
-  const { roles } = useUserAccessStateContext();
-  const canRunChecks = roles.some((role) => bulkRunAllowedRoles.includes(role));
-
-  const toggleCheck = useCallback((id: string) => {
-    setSelectedCheckIds((prev) => {
-      const next = new Set(prev);
-      if (next.has(id)) {
-        next.delete(id);
-      } else {
-        next.add(id);
-      }
-      return next;
-    });
-  }, []);
-
-  const toggleChecks = useCallback((ids: string[], selected: boolean) => {
-    setSelectedCheckIds((prev) => {
-      const next = new Set(prev);
-      ids.forEach((id) => {
-        if (selected) {
-          next.add(id);
-        } else {
-          next.delete(id);
-        }
-      });
-      return next;
-    });
-  }, []);
-
-  const exitBulkMode = useCallback(() => {
-    setIsBulkRunMode(false);
-    setSelectedCheckIds(new Set());
-  }, []);
-
-  const bulkRunContextValue = useMemo(
-    () => ({
-      isBulkRunMode,
-      selectedCheckIds,
-      toggleCheck,
-      toggleChecks
-    }),
-    [isBulkRunMode, selectedCheckIds, toggleCheck, toggleChecks]
-  );
+  const { isBulkRunMode, exitBulkMode } = useBulkCheckRun();
 
   const {
     healthState: { checks, filteredChecks },
@@ -238,65 +186,53 @@ export function Canary({
   }
 
   return (
-    <BulkCheckRunContext.Provider value={bulkRunContextValue}>
-      <div
-        className={clsx(
-          "flex w-full flex-row place-content-center",
-          isCanaryUI
-            ? "h-screen min-w-[800px] overflow-auto"
-            : "h-full overflow-y-auto"
-        )}
-      >
-        <CanarySidebar
-          isMenuItemOpen={isMenuItemOpen}
-          setIsMenuItemOpen={setIsMenuItemOpen}
-        />
+    <div
+      className={clsx(
+        "flex w-full flex-row place-content-center",
+        isCanaryUI
+          ? "h-screen min-w-[800px] overflow-auto"
+          : "h-full overflow-y-auto"
+      )}
+    >
+      <CanarySidebar
+        isMenuItemOpen={isMenuItemOpen}
+        setIsMenuItemOpen={setIsMenuItemOpen}
+      />
 
-        <div className="flex h-full flex-1 flex-col overflow-y-auto p-6">
-          <div className="mb-2 flex flex-wrap items-center gap-2">
-            <div
-              role="button"
-              title="Toggle filter menu"
-              onClick={() => {
-                setIsMenuItemOpen((prev) => !prev);
-              }}
-              className="relative flex cursor-pointer items-center gap-1 px-4 2xl:hidden"
-            >
-              <FaFilter />
-              <Badge text={totalLabelsApplied} />
-            </div>
-
-            <CanaryFiltersBar checks={checks ?? []} />
-
-            {canRunChecks && !isBulkRunMode && (
-              <Button
-                className="btn-white"
-                size="sm"
-                icon={<VscDebugRerun />}
-                text="Bulk Run"
-                onClick={() => setIsBulkRunMode(true)}
-              />
-            )}
+      <div className="flex h-full flex-1 flex-col overflow-y-auto p-6">
+        <div className="mb-2 flex flex-wrap items-center gap-2">
+          <div
+            role="button"
+            title="Toggle filter menu"
+            onClick={() => {
+              setIsMenuItemOpen((prev) => !prev);
+            }}
+            className="relative flex cursor-pointer items-center gap-1 px-4 2xl:hidden"
+          >
+            <FaFilter />
+            <Badge text={totalLabelsApplied} />
           </div>
 
-          {isBulkRunMode && (
-            <div className="mb-2">
-              <BulkCheckRunBar
-                filteredChecks={filteredChecks}
-                onExit={exitBulkMode}
-              />
-            </div>
-          )}
+          <CanaryFiltersBar checks={checks ?? []} />
+        </div>
 
-          <div className="flex flex-1 flex-col pb-4">
-            <CanaryInterfaceMinimal
-              checks={checks ?? undefined}
-              onLabelFiltersCallback={labelUpdateCallback}
-              onFilterCallback={updateFilteredChecks}
+        {isBulkRunMode && (
+          <div className="mb-2">
+            <BulkCheckRunBar
+              filteredChecks={filteredChecks}
+              onExit={exitBulkMode}
             />
           </div>
+        )}
+
+        <div className="flex flex-1 flex-col pb-4">
+          <CanaryInterfaceMinimal
+            checks={checks ?? undefined}
+            onLabelFiltersCallback={labelUpdateCallback}
+            onFilterCallback={updateFilteredChecks}
+          />
         </div>
       </div>
-    </BulkCheckRunContext.Provider>
+    </div>
   );
 }
